Implemented authorization

+++ InnoGotchiWebAPI/Controllers/InnoGotchiPetsController.cs
@@ -1,12 +1,9 @@
 using InnoGotchiWebAPI.Interfaces;
 using InnoGotchiWebAPI.Logic;
 using InnoGotchiWebAPI.Models;
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
-using System.Text.Json;
+using System.Security.Claims;
 
 namespace InnoGotchiWebAPI.Controllers
 {
@@ -34,19 +31,21 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pets succesfully retrieved</response>
-        [Authorize]
+        /// <response code="401">Unauthorized</response>
         [HttpGet]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
         public async Task<ActionResult<IEnumerable<ClientPetModel>>> GetPets()
         {
-            var dbPets = await _dbService.GetPets();
+            var userlogin = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            var dbPets = await _dbService.GetPets(userlogin!);
 
             var clientPets = Mappers.PetDbToClientMapper.Map<IEnumerable<ClientPetModel>>(dbPets);
 
             return new(clientPets);
         }
-
-
 
         /// <summary>
         /// Get pet with given id from db
@@ -60,13 +59,23 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully retrieved</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpGet("{id:Guid}")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         public async Task<ActionResult<ClientPetModel>> GetPet(Guid id)
         {
             var dbPet = await _dbService.GetPet(id);
+
+            if (CheckOwnership(dbPet) == false)
+            {
+                return Forbid();
+            }
 
             var clientPet = Mappers.PetDbToClientMapper.Map<ClientPetModel>(dbPet);
 
@@ -84,13 +93,26 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully deleted</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpDelete("{id:Guid}")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         public async Task<ActionResult<ClientPetModel>> DeletePet(Guid id)
         {
-            var dbPet = await _dbService.DeletePet(id);
+            // todo: Is this optimal?
+            var dbPet = await _dbService.GetPet(id);
+
+            if (CheckOwnership(dbPet) == false)
+            {
+                return Forbid();
+            }
+
+            dbPet = await _dbService.DeletePet(id);
 
             var clientPet = Mappers.PetDbToClientMapper.Map<ClientPetModel>(dbPet);
 
@@ -108,15 +130,23 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully created</response>
+        /// <response code="401">Unauthorized</response>
         /// <response code="409">Pet with given id already exists. Use PUT instead</response>
         [HttpPost]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
         [ProducesResponseType(409)]
         public async Task<ActionResult> PostPet(ClientPetModel pet)
         {
             var dbPet = Mappers.PetClientToDbMapper.Map<DbPetModel>(pet);
 
-            await _dbService.PostPet(dbPet);
+            var userlogin = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            // Todo: is it ok to set ownership here?
+            dbPet.OwnerId = userlogin!;
+
+            await _dbService.PostPet(dbPet, userlogin!);
 
             return Ok();
         }
@@ -132,12 +162,29 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully updated</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpPut]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         public async Task<ActionResult> PutPet(ClientPetModel pet)
         {
+            bool petExists = await _dbService.PetExists(pet.Id);
+
+            if (petExists)
+            {
+                var ownerCheckingPet = await _dbService.GetPet(pet.Id);
+
+                if (CheckOwnership(ownerCheckingPet) == false)
+                {
+                    return Forbid();
+                }
+            }
+
             var dbPet = Mappers.PetClientToDbMapper.Map<DbPetModel>(pet);
 
             await _dbService.PutPet(dbPet);
@@ -156,15 +203,27 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully updated</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpGet("{id:Guid}/feed")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         public async Task<ActionResult<ClientPetModel>> Feed(Guid id)
         {
+            var dbPet = await _dbService.GetPet(id);
+
+            if (CheckOwnership(dbPet) == false)
+            {
+                return Forbid();
+            }
+
             var pet = await _petUpdateService.Feed(id);
 
-            var clientPet = Mappers.PetClientToDbMapper.Map<ClientPetModel>(pet);
+            var clientPet = Mappers.PetDbToClientMapper.Map<ClientPetModel>(pet);
 
             return new(clientPet);
         }
@@ -180,15 +239,27 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully updated</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpGet("{id:Guid}/givedrink")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         public async Task<ActionResult<ClientPetModel>> GiveDrink(Guid id)
         {
+            var dbPet = await _dbService.GetPet(id);
+
+            if (CheckOwnership(dbPet) == false)
+            {
+                return Forbid();
+            }
+
             var pet = await _petUpdateService.GiveDrink(id);
 
-            var clientPet = Mappers.PetClientToDbMapper.Map<ClientPetModel>(pet);
+            var clientPet = Mappers.PetDbToClientMapper.Map<ClientPetModel>(pet);
 
             return new(clientPet);
         }
@@ -204,17 +275,29 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pet succesfully updated</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Pet belongs to other user</response>
         /// <response code="404">Pet with given id not found</response>
         [HttpGet("{id:Guid}/update")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
-        public async Task<ActionResult<DbPetModel>> Update(Guid id)
+        public async Task<ActionResult<ClientPetModel>> Update(Guid id)
         {
+            var dbPet = await _dbService.GetPet(id);
+
+            if (CheckOwnership(dbPet) == false)
+            {
+                return Forbid();
+            }
+
             var pet = await _petUpdateService.Update(id);
 
-            // map
+            var clientPet = Mappers.PetDbToClientMapper.Map<ClientPetModel>(pet);
 
-            return new(pet);
+            return new(clientPet);
         }
 
         /// <summary>
@@ -227,13 +310,32 @@ namespace InnoGotchiWebAPI.Controllers
         ///
         /// </remarks>
         /// <response code="200">Pets succesfully updated</response>
+        /// <response code="401">Unauthorized</response>
         [HttpGet("update")]
+        [Authorize]
         [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
         public async Task<ActionResult> UpdateAll()
         {
-            await _petUpdateService.UpdateAll();
+            var userlogin = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            await _petUpdateService.UpdateAll(userlogin!);
 
             return Ok();
+        }
+
+        // todo: Bad practice? How to implement check properly?
+        [NonAction]
+        private bool CheckOwnership(DbPetModel pet)
+        {
+            var userlogin = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (pet.OwnerId != userlogin)
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
